Move HTTPRequest.url to the main library
